requirements.txt
phidata
duckduckgo-search
groq
yfinance
duckduckgo-search
fastapi
python-multipart
uvicorn
